+++ src/models/start_number.rs
@@ -1,8 +1,8 @@
 use diesel::prelude::*;
-use diesel::sql_types::Int4;
+use diesel::sql_types::BigInt;
 
 #[derive(QueryableByName)]
 pub struct StartNumber {
-    #[diesel(sql_type = Int4)]
-    pub start_number: i32
+    #[diesel(sql_type = BigInt)]
+    pub start_number: i64
 }
